+++ src/main/java/ladder/controller/Controller.java
@@ -5,14 +5,14 @@ import static ladder.view.InputView.QUIT_RESULT_COMMAND;
 
 import java.util.List;
 import ladder.domain.DirectionGenerator;
-import ladder.domain.Height;
-import ladder.domain.Ladder;
-import ladder.domain.Players;
 import ladder.domain.RandomDirectionGenerator;
-import ladder.domain.Result;
-import ladder.domain.Results;
 import ladder.domain.ResultsOfPlayers;
-import ladder.domain.Width;
+import ladder.domain.ladder.Height;
+import ladder.domain.ladder.Ladder;
+import ladder.domain.ladder.Width;
+import ladder.domain.player.Players;
+import ladder.domain.result.Result;
+import ladder.domain.result.Results;
 import ladder.exception.ExceptionHandler;
 import ladder.view.InputView;
 import ladder.view.OutputView;
