+++ src/main/java/ladder/domain/DirectionGenerator.java
@@ -1,5 +1,7 @@
 package ladder.domain;
 
+import ladder.domain.ladder.Direction;
+
 public interface DirectionGenerator {
     Direction generate();
 }
