+++ src/main/java/ladder/domain/RandomDirectionGenerator.java
@@ -1,9 +1,10 @@
 package ladder.domain;
 
-import static ladder.domain.Direction.NONE;
-import static ladder.domain.Direction.RIGHT;
+import static ladder.domain.ladder.Direction.NONE;
+import static ladder.domain.ladder.Direction.RIGHT;
 
 import java.util.Random;
+import ladder.domain.ladder.Direction;
 
 public class RandomDirectionGenerator implements DirectionGenerator {
     private static final Random RANDOM = new Random();
