+++ src/main/java/ladder/domain/ladder/Direction.java
@@ -1,4 +1,6 @@
-package ladder.domain;
+package ladder.domain.ladder;
+
+import ladder.domain.DirectionGenerator;
 
 public enum Direction {
     LEFT(-1),
