+++ src/main/java/ladder/domain/ladder/Width.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.List;
 import java.util.function.Supplier;
