+++ src/test/java/ladder/domain/DirectionTest.java
@@ -1,11 +1,12 @@
 package ladder.domain;
 
-import static ladder.domain.Direction.LEFT;
-import static ladder.domain.Direction.NONE;
-import static ladder.domain.Direction.RIGHT;
+import static ladder.domain.ladder.Direction.LEFT;
+import static ladder.domain.ladder.Direction.NONE;
+import static ladder.domain.ladder.Direction.RIGHT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import ladder.domain.ladder.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
