+++ src/test/java/ladder/domain/LocationTest.java
@@ -3,6 +3,7 @@ package ladder.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import ladder.domain.player.Location;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
