+++ src/test/java/ladder/domain/ResultsTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.stream.Stream;
+import ladder.domain.player.Location;
+import ladder.domain.result.Result;
+import ladder.domain.result.Results;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
